backend/requirements.txt
fastapi
uvicorn[standard]
pymysql
sqlalchemy
redis
pydantic
python-dotenv
langchain
langchain-openai
chromadb
sentence-transformers
kafka-python
